+++ docs/src/components/Sidebar/Sidebar.tsx
@@ -276,6 +276,14 @@ class Sidebar extends React.Component<any, any> {
                   <Menu.Item
                     as={NavLink}
                     exact
+                    to="/prototype-async-dropdown-search"
+                    activeClassName="active"
+                  >
+                    Async Dropdown Search
+                  </Menu.Item>
+                  <Menu.Item
+                    as={NavLink}
+                    exact
                     to="/prototype-async-shorthand"
                     activeClassName="active"
                   >