+++ docs/src/examples/components/Dropdown/index.tsx
@@ -2,12 +2,14 @@ import * as React from 'react'
 
 import Types from './Types'
 import Variations from './Variations'
+import State from './State'
 import Usage from './Usage'
 
 const DropdownExamples = () => (
   <div>
     <Types />
     <Variations />
+    <State />
     <Usage />
   </div>
 )