+++ docs/src/prototypes/chatPaneHeader/chatTitle.tsx
@@ -4,11 +4,9 @@ import {
   Popup,
   Menu,
   popupFocusTrapBehavior,
-  Avatar,
-  Icon,
-  Header,
-  Box,
   Input,
+  Avatar,
+  Header,
   navigableListBehavior,
   navigableListItemBehavior,
 } from '@stardust-ui/react'
@@ -24,21 +22,31 @@ const menuStyles = ({ theme: { siteVariables } }) => ({
   marginTop: '5px',
 })
 
-const PopupItemLayout = props => {
-  return <Popup trigger={<Box {...props} />} content="Hello from popup!" />
+const MoreParticipantsButton = props => {
+  const { count, ...rest } = props
+  return (
+    <Button
+      role="listitem"
+      {...rest}
+      circular
+      aria-label={`${count} participants`}
+      aria-haspopup="true"
+      aria-posinset={count}
+      aria-setsize={count}
+    >{`+${count}`}</Button>
+  )
 }
 
 class ChatTitle extends React.Component<any> {
-  private getButtonWithAllParticipants(listItems: any) {
+  private getButtonWithAllParticipants(verticalItems: any) {
     const MenuButton = props => (
       <Button
         {...props}
         circular
-        aria-label={`${listItems.length} participants`}
+        aria-label={`${verticalItems.length} participants`}
         aria-haspopup="true"
-      >{`+${listItems.length}`}</Button>
+      >{`+${verticalItems.length}`}</Button>
     )
-
     const items = [
       {
         styles: {
@@ -52,7 +60,7 @@ class ChatTitle extends React.Component<any> {
         menu: {
           styles: menuStyles,
           pills: true,
-          items: this.getMenuItemsAsPopupTrigger(listItems),
+          items: verticalItems,
         },
       },
     ]
@@ -68,7 +76,60 @@ class ChatTitle extends React.Component<any> {
     )
   }
 
-  private userListFromMenu(userList): any {
+  private createHorizontalItem = user => {
+    const item = {
+      as: 'li',
+      role: 'listitem',
+      key: user.key,
+      content: (
+        <>
+          {' '}
+          {user.media} <span> {user.content} </span>{' '}
+        </>
+      ),
+      userName: user.content,
+      'aria-haspopup': 'dialog',
+      accessibility: navigableListItemBehavior,
+      wrapper: null,
+    }
+
+    return render => render(item, this.renderUserWithCard)
+  }
+
+  private createVerticalItem = user => {
+    const item = {
+      key: user.key,
+      content: user.content,
+      'aria-haspopup': 'dialog',
+      icon: (
+        <Avatar
+          image={{
+            src: 'public/images/avatar/small/matt.jpg',
+          }}
+          status={{ color: 'green', icon: 'check', title: 'Available' }}
+        />
+      ),
+    }
+
+    return render => render(item, this.renderUserWithCard)
+  }
+
+  private createOverflow = menuItems => ({
+    key: 'restOfParticipants',
+    'aria-haspopup': true,
+    'aria-label': `${menuItems.length} more participants`,
+    accessibility: navigableListItemBehavior,
+    wrapper: { as: 'li', role: 'presentation' },
+    as: MoreParticipantsButton,
+    count: menuItems.length,
+    menu: {
+      styles: menuStyles,
+      pills: true,
+      items: menuItems,
+    },
+  })
+
+  private renderHorizontalMenu = items => {
     return (
       <Menu
         variables={{
@@ -81,129 +142,55 @@ class ChatTitle extends React.Component<any> {
         defaultActiveIndex={0}
         accessibility={navigableListBehavior}
         aria-label="chat participants"
-        items={this.getItemsForMenuAsList(userList)}
+        items={items}
       />
     )
   }
 
-  getMenuItemsAsPopupTrigger(items): any {
-    const newMenuItems = []
-    items.map(listItem => {
-      newMenuItems.push({
-        'aria-haspopup': 'dialog',
-        key: listItem.key,
-        content: listItem.content,
-        icon: (
-          <Avatar
-            image={{
-              src: 'public/images/avatar/small/matt.jpg',
-            }}
-            status={{ color: 'green', icon: 'check', title: 'Available' }}
-          />
-        ),
-      })
-    })
-
-    const leaveChat = {
-      onClick: e => window.alert('leave chat popup will be here'),
-      key: 'leaveChat',
-      content: 'Leave chat',
-      icon: <Icon name="leave" />,
-    }
-
-    const newMenuItemsWithLeave = newMenuItems.map(item => render => render(item, this.renderPopup))
-    newMenuItemsWithLeave.push(render => render(leaveChat))
-    return newMenuItemsWithLeave
-  }
-
-  private getItemsForMenuAsList(userList): any {
-    const newMenuAsListItems = []
-    const MenuButton = props => (
-      <Button
-        {...props}
-        circular
-        aria-label={`${newMenuAsListItems.length} more participants`}
-        aria-haspopup="true"
-      >{`+${newMenuAsListItems.length}`}</Button>
-    )
-
-    userList.slice(0, 3).map(user => {
-      newMenuAsListItems.push({
-        as: PopupItemLayout,
-        key: user.key,
-        content: (
-          <>
-            {user.media} <span> {user.content} </span>
-          </>
-        ),
-        role: 'button',
-        accessibility: navigableListItemBehavior,
-      })
-    })
-
-    if (userList.length > 3) {
-      newMenuAsListItems.push({
-        key: 'restOfParticipants',
-        role: 'button',
-        'aria-haspopup': true,
-        'aria-label': `${newMenuAsListItems.length} more participants`,
-        accessibility: navigableListItemBehavior,
-        as: MenuButton,
-        menu: {
-          styles: menuStyles,
-          pills: true,
-          items: this.getMenuItemsAsPopupTrigger(userList.slice(3, userList.length)),
-        },
-      })
-    }
-    return newMenuAsListItems
-  }
-
-  private renderPopup = (MenuItem, props) => {
+  private renderUserWithCard = (MenuItem, props) => {
+    const { userName, ...rest } = props
     return (
       <Popup
-        styles={{ marginLeft: '10px', backgroundColor: 'white' }}
-        on="hover"
-        key={props.key}
-        position="after"
-        align="bottom"
         // accessibility={popupFocusTrapBehavior}
-        trigger={<MenuItem {...props} />}
-        content={
-          <div>
-            <Header as="h4">Any content.</Header>
-            <Input icon="search" placeholder="Search..." />
-            <Button>Testing button</Button>
-            <Input icon="search" placeholder="Search..." />
-            <Button>Testing button</Button>
-            <Input icon="search" placeholder="Search..." />
-            <Button>Testing button</Button>
-          </div>
-        }
+        aria-haspopup="dialog"
+        on="hover"
+        trigger={<MenuItem {...rest} />}
+        content={{
+          content: (
+            <>
+              <Header as="h4">{userName} card</Header>
+              <Input placeholder="Search..." />
+              <Button icon="send" title="Send message" />
+            </>
+          ),
+        }}
       />
     )
-  }
-
-  private renderTitleOrUserList(userList, groupChatName): any {
-    if (groupChatName) {
-      return (
-        <div style={{ display: 'flex' }}>
-          <Header styles={headingStyle} as="h2" content={groupChatName} className="no-anchor" />
-          {this.getButtonWithAllParticipants(userList)}
-        </div>
-      )
-    }
-    if (userList.length <= 3) {
-      return this.userListFromMenu(userList)
-    }
-    if (userList.length > 3) {
-      return <div style={{ display: 'flex' }}>{this.userListFromMenu(userList)}</div>
-    }
   }
 
   public render() {
     const { listItems, groupChatName } = this.props
-    return this.renderTitleOrUserList(listItems, groupChatName)
+    const horizontalUsers = listItems.slice(0, 3)
+
+    if (groupChatName) {
+      const verticalUsers = listItems.map(user => this.createVerticalItem(user))
+      return (
+        <div style={{ display: 'flex' }}>
+          <Header styles={headingStyle} as="h2" content={groupChatName} className="no-anchor" />
+          {this.getButtonWithAllParticipants(verticalUsers)}
+        </div>
+      )
+    }
+
+    const verticalUsers = listItems.slice(3, listItems.length)
+
+    const horizontalItems = horizontalUsers.map(user => this.createHorizontalItem(user))
+    const verticalItems = verticalUsers.map(user => this.createVerticalItem(user))
+
+    if (verticalItems.length > 0) {
+      horizontalItems.push(this.createOverflow(verticalItems))
+    }
+    return this.renderHorizontalMenu(horizontalItems)
   }
 }
 export default ChatTitle